Crate gstreamer_net

Modules

prelude

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

NetClientClock
NetClientClock implements a custom gst::Clock that synchronizes its time to a remote time provider such as NetTimeProvider. NtpClock implements a gst::Clock that synchronizes its time to a remote NTPv4 server.

NetTimeProvider
This object exposes the time of a gst::Clock on the network.

NtpClock
Implements

NetClientClockExt, gst::ClockExt, gst::ObjectExt, glib::object::ObjectExt

PtpClock
GstPtpClock implements a PTP (IEEE1588:2008) ordinary clock in slave-only mode, that allows a GStreamer pipeline to synchronize to a PTP network clock in some specific domain.

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.