Struct gstreamer_net::NetClientClock

pub struct NetClientClock(_, _);

NetClientClock implements a custom gst::Clock that synchronizes its time to a remote time provider such as NetTimeProvider. NtpClock implements a gst::Clock that synchronizes its time to a remote NTPv4 server.

A new clock is created with NetClientClock::new or NtpClock::new, which takes the address and port of the remote time provider along with a name and an initial time.

This clock will poll the time provider and will update its calibration parameters based on the local and remote observations.

The "round-trip" property limits the maximum round trip packets can take.

Various parameters of the clock can be configured with the parent gst::Clock "timeout", "window-size" and "window-threshold" object properties.

A NetClientClock and NtpClock is typically set on a gst::Pipeline with gst::Pipeline::use_clock.

If you set a gst::Bus on the clock via the "bus" object property, it will send gst::MessageType::Element messages with an attached gst::Structure containing statistics about clock accuracy and network traffic.

Implements

gst::ClockExt, gst::ObjectExt, glib::object::ObjectExt

Methods

impl NetClientClock

pub fn get_property_address(&self) -> Option<String>

pub fn set_property_address(&self, address: Option<&str>)

pub fn get_property_base_time(&self) -> u64

pub fn get_property_bus(&self) -> Option<Bus>

pub fn set_property_bus(&self, bus: Option<&Bus>)

pub fn get_property_internal_clock(&self) -> Option<Clock>

pub fn get_property_minimum_update_interval(&self) -> u64

pub fn set_property_minimum_update_interval(&self, minimum_update_interval: u64)

pub fn get_property_port(&self) -> i32

pub fn set_property_port(&self, port: i32)

pub fn get_property_qos_dscp(&self) -> i32

pub fn set_property_qos_dscp(&self, qos_dscp: i32)

pub fn get_property_round_trip_limit(&self) -> u64

pub fn set_property_round_trip_limit(&self, round_trip_limit: u64)

pub fn connect_property_address_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_base_time_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_bus_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_internal_clock_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_minimum_update_interval_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_port_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_qos_dscp_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_round_trip_limit_notify<F: Fn(&NetClientClock) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl NetClientClock

pub fn new<'a, P: Into<Option<&'a str>>>(
name: P,
remote_address: &str,
remote_port: i32,
base_time: ClockTime
) -> NetClientClock

Create a new GstNetClientInternalClock that will report the time provided by the NetTimeProvider on remote_address and remote_port.

name

a name for the clock

remote_address

the address or hostname of the remote clock provider

remote_port

the port of the remote clock provider

base_time

initial time of the clock

Returns

a new gst::Clock that receives a time from the remote clock.