Struct gstreamer_net::NetTimeProvider

pub struct NetTimeProvider(_, _);

This object exposes the time of a gst::Clock on the network.

A NetTimeProvider is created with NetTimeProvider::new which takes a gst::Clock, an address and a port number as arguments.

After creating the object, a client clock such as NetClientClock can query the exposed clock over the network for its values.

The NetTimeProvider typically wraps the clock used by a gst::Pipeline.

Implements

gst::ObjectExt, glib::object::ObjectExt

Methods

impl NetTimeProvider

pub fn get_property_active(&self) -> bool

pub fn set_property_active(&self, active: bool)

pub fn get_property_address(&self) -> Option<String>

pub fn get_property_clock(&self) -> Option<Clock>

pub fn get_property_port(&self) -> i32

pub fn get_property_qos_dscp(&self) -> i32

pub fn set_property_qos_dscp(&self, qos_dscp: i32)

pub fn connect_property_active_notify<F: Fn(&NetTimeProvider) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_address_notify<F: Fn(&NetTimeProvider) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_clock_notify<F: Fn(&NetTimeProvider) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_port_notify<F: Fn(&NetTimeProvider) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_qos_dscp_notify<F: Fn(&NetTimeProvider) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl NetTimeProvider

pub fn new<'a, P: IsA<Clock>, Q: Into<Option<&'a str>>>(
clock: &P,
address: Q,
port: i32
) -> NetTimeProvider

Allows network clients to get the current time of clock.

clock

a gst::Clock to export over the network

address

an address to bind on as a dotted quad (xxx.xxx.xxx.xxx), IPv6 address, or NULL to bind to all addresses

port

a port to bind on, or 0 to let the kernel choose

Returns

the new NetTimeProvider, or NULL on error